requirements.txt
langgraph
langgraph-cli
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
langchain-experimental
notebook
python-dotenv
langchain-chroma
azure-identity
scikit-learn
openevals
openai
ipython
pyppeteer
langgraph-supervisor
langgraph-cli[inmem]
langchain-anthropic
langchain-google-vertexai
openevals
langgraph-swarm
langchain-aws
playwright